requirements.txt
attrs==21.4.0
Faker==13.3.1
falcon==3.0.1
flake8==4.0.1
gunicorn==20.1.0
importlib-resources==5.4.0
iniconfig==1.1.1
jsonschema==4.4.0
mccabe==0.6.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.0.1
python-dateutil==2.8.2
python-mimeparse==1.6.0
six==1.16.0
tomli==2.0.1
zipp==3.7.0
